privtangle shutdown, while keeping state

+++ packages/testutil/privtangle/privtangle.go
@@ -79,21 +79,21 @@ func Start(ctx context.Context, baseDir string, basePort, nodeCount int, logfunc
 	}
 
 	pt.generateSnapshot()
-	pt.StartServers()
+	pt.StartServers(true)
 
 	return &pt
 }
 
-func (pt *PrivTangle) StartServers() {
+func (pt *PrivTangle) StartServers(deleteExisting bool) {
 	for i := range pt.NodeKeyPairs {
-		pt.startNode(i)
+		pt.startNode(i, deleteExisting)
 		time.Sleep(500 * time.Millisecond) // TODO: Remove?
 	}
 	pt.logf("Starting... all nodes started.")
 
 	pt.waitAllReady(20 * time.Second)
 	pt.logf("Starting... all nodes are up and running, starting coordinator.")
-	pt.startCoordinator(0)
+	pt.startCoordinator(0, deleteExisting)
 
 	pt.waitAllHealthy(20 * time.Second)
 	pt.logf("Starting... coordinator started, all nodes are healthy.")
@@ -133,7 +133,7 @@ func (pt *PrivTangle) generateSnapshot() {
 	}
 }
 
-func (pt *PrivTangle) startNode(i int) {
+func (pt *PrivTangle) startNode(i int, deleteExisting bool) {
 	env := []string{}
 	nodePath := filepath.Join(pt.BaseDir, fmt.Sprintf("node-%d", i))
 	nodePathDB := "db"               // Relative from nodePath.
@@ -142,28 +142,30 @@ func (pt *PrivTangle) startNode(i int) {
 	nodePathSnapFull := fmt.Sprintf("%s/full_snapshot.bin", nodePathSnapshots)
 	nodePathSnapDelta := fmt.Sprintf("%s/delta_snapshot.bin", nodePathSnapshots)
 
-	if err := os.RemoveAll(nodePath); err != nil {
-		panic(fmt.Errorf("unable to delete dir %v: %w", nodePath, err))
-	}
-	if err := os.MkdirAll(nodePath, 0o755); err != nil {
-		panic(fmt.Errorf("unable to create dir %v: %w", nodePath, err))
-	}
-	if err := os.MkdirAll(filepath.Join(nodePath, nodePathDB), 0o755); err != nil {
-		panic(fmt.Errorf("unable to create dir %v: %w", nodePathDB, err))
-	}
-	if err := os.MkdirAll(filepath.Join(nodePath, nodePathSnapshots), 0o755); err != nil {
-		panic(fmt.Errorf("unable to create dir %v: %w", nodePathSnapshots, err))
-	}
-	if err := os.WriteFile(filepath.Join(nodePath, pt.ConfigFile), []byte(pt.configFileContent()), 0o600); err != nil {
-		panic(fmt.Errorf("unable to create %s: %w", pt.ConfigFile, err))
-	}
+	if deleteExisting {
+		if err := os.RemoveAll(nodePath); err != nil {
+			panic(fmt.Errorf("unable to delete dir %v: %w", nodePath, err))
+		}
+		if err := os.MkdirAll(nodePath, 0o755); err != nil {
+			panic(fmt.Errorf("unable to create dir %v: %w", nodePath, err))
+		}
+		if err := os.MkdirAll(filepath.Join(nodePath, nodePathDB), 0o755); err != nil {
+			panic(fmt.Errorf("unable to create dir %v: %w", nodePathDB, err))
+		}
+		if err := os.MkdirAll(filepath.Join(nodePath, nodePathSnapshots), 0o755); err != nil {
+			panic(fmt.Errorf("unable to create dir %v: %w", nodePathSnapshots, err))
+		}
+		if err := os.WriteFile(filepath.Join(nodePath, pt.ConfigFile), []byte(pt.configFileContent()), 0o600); err != nil {
+			panic(fmt.Errorf("unable to create %s: %w", pt.ConfigFile, err))
+		}
 
-	snapContents, err := os.ReadFile(filepath.Join(pt.BaseDir, pt.SnapshotInit))
-	if err != nil {
-		panic(fmt.Errorf("unable to read initial snapshot : %w", err))
-	}
-	if err := os.WriteFile(filepath.Join(nodePath, nodePathSnapFull), snapContents, 0o600); err != nil {
-		panic(fmt.Errorf("unable to copy the initial snapshot : %w", err))
+		snapContents, err := os.ReadFile(filepath.Join(pt.BaseDir, pt.SnapshotInit))
+		if err != nil {
+			panic(fmt.Errorf("unable to read initial snapshot : %w", err))
+		}
+		if err := os.WriteFile(filepath.Join(nodePath, nodePathSnapFull), snapContents, 0o600); err != nil {
+			panic(fmt.Errorf("unable to copy the initial snapshot : %w", err))
+		}
 	}
 
 	args := []string{
@@ -203,7 +205,7 @@ func (pt *PrivTangle) startNode(i int) {
 	}
 }
 
-func (pt *PrivTangle) startCoordinator(i int) *exec.Cmd {
+func (pt *PrivTangle) startCoordinator(i int, deleteExisting bool) *exec.Cmd {
 	env := []string{
 		fmt.Sprintf("COO_PRV_KEYS=%s,%s",
 			hex.EncodeToString(pt.CooKeyPair1.GetPrivateKey().AsBytes()),
@@ -211,13 +213,15 @@ func (pt *PrivTangle) startCoordinator(i int) *exec.Cmd {
 		),
 	}
 	args := []string{
-		"--cooBootstrap",
 		"--cooStartIndex=0",
 		"--coordinator.interval=100ms",
 		"--coordinator.debugFakeMilestoneTimestamps=true",
 		// no need to keep a backup of the coo milestones
 		"--coordinator.blockBackups.enabled=false",
 		fmt.Sprintf("--inx.address=0.0.0.0:%d", pt.NodePortINX(i)),
+	}
+	if deleteExisting {
+		args = append(args, "--cooBootstrap")
 	}
 	return pt.startINXPlugin(i, "inx-coordinator", args, env)
 }
